backend/api_notificaciones/requirements.txt
flask
flask-cors
requests
python-dotenv
twilio
pywhatkit